refactor(replay-store): move detectReplayFormat to shared for headless reuse
detectReplayFormat lived in electron/session-io.ts, which imports electron.
Move it next to parseReplay/serializeEvents in shared/replay-store.ts so the
headless replay runner can sniff replay-vs-transcript without pulling in the
electron module, and re-export it from session-io.ts to preserve the existing
import surface. No behavior change.
